+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.16.2] - 2024-08-05
+
 ### Added
  - Add the experimental `gdeflate` bytes-to-bytes codec
  - Add `Array::chunk_key()`
@@ -922,7 +924,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Initial public release
 
-[unreleased]: https://github.com/LDeakin/zarrs/compare/v0.16.1...HEAD
+[unreleased]: https://github.com/LDeakin/zarrs/compare/v0.16.2...HEAD
+[0.16.2]: https://github.com/LDeakin/zarrs/releases/tag/v0.16.2
 [0.16.1]: https://github.com/LDeakin/zarrs/releases/tag/v0.16.1
 [0.16.0]: https://github.com/LDeakin/zarrs/releases/tag/v0.16.0
 [0.15.1]: https://github.com/LDeakin/zarrs/releases/tag/v0.15.1